scanner move
