+++ src/jca/springframework/controller/MappingBuilder.java
@@ -4,10 +4,10 @@ import java.lang.reflect.Method;
 import java.util.HashMap;
 import java.util.List;
 
-import jca.springframework.MethodScanner;
-import jca.springframework.PackageScanner;
 import jca.springframework.annotations.Controller;
 import jca.springframework.annotations.Get;
+import jca.springframework.scanner.MethodScanner;
+import jca.springframework.scanner.PackageScanner;
 
 public class MappingBuilder {
